+++ LevelSolverThread.cpp
@@ -1,4 +1,5 @@
 #include "LevelSolverThread.h"
+#include "DriveLedger.h"
 #include "MergeFiles.h"
 #include "BlasterFile.h"
 #include "BlasterFileName.h"